adjust input token ratio calculation method

+++ test/forkMainnet/SmartOrderStrategy/Setup.t.sol
@@ -14,7 +14,7 @@ contract SmartOrderStrategyTest is Test, Tokens, BalanceUtil {
     address defaultInputToken = USDC_ADDRESS;
     address defaultOutputToken = WETH_ADDRESS;
     uint256 defaultInputAmount = 1000;
-    uint128 defaultInputRatio = 5000;
+    uint256 defaultInputRatio = 5000;
     uint256 defaultExpiry = block.timestamp + 100;
     bytes defaultOpsData;
     bytes encodedUniv3Path;
